move autocomplete component
