+++ src/components/AutoCompleteMajor.tsx
