+++ src/components/onboarding/welcome.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { SemesterCode } from '@prisma/client';
 import { MenuItem, Select, SelectChangeEvent } from '@mui/material';
 import EmojiIcons from '@/icons/EmojiIcon';
-import AutoCompleteMajor from '@/pages/auth/AutoCompleteMajor';
+import AutoCompleteMajor from '@/components/AutoCompleteMajor';
 import majorsList from '@data/majors.json';
 import useSearch from '../search/search';
